libcex: move the BPF FFI
Pure move of the classic-BPF instruction (BpfInsn, cross-platform) into libcex::bpf, and the BSD BPF-device items (DLT_EN10MB/DLT_NULL, the bpf_program struct, BPF_ALIGN, and the BPF_WORDALIGN reimpl) into a BSD-gated libcex::bpf_device. Only visibility widened to pub(crate); the per-OS anchoring asserts and the wordalign test travel with their items. capture/{bpf,filter,af_packet}.rs now reference libcex::.
